+++ apps/cli/cli_main.c
@@ -552,7 +552,7 @@ main(int argc, char **argv)
 	free(restarg);
     // Gets in your face if we log on stderr
     clicon_log_init(__PROGRAM__, LOG_INFO, 0); /* Log on syslog no stderr */
-    clicon_log(LOG_NOTICE, "%s: %u Terminated\n", __PROGRAM__, getpid());
+    clicon_log(LOG_NOTICE, "%s: %u Terminated", __PROGRAM__, getpid());
     if (h)
 	cli_terminate(h);
     return retval;