+++ lib/clixon/clixon_yang_module.h
@@ -52,6 +52,6 @@
  */
 int yang_modules_init(clicon_handle h);
 char *yang_modules_revision(clicon_handle h);
-int yang_modules_state_get(clicon_handle h, yang_spec *yspec, cxobj **xret);
+int yang_modules_state_get(clicon_handle h, yang_spec *yspec, int brief, cxobj **xret);
 
 #endif  /* _CLIXON_YANG_MODULE_H_ */